GC: Created gc logs and pdf reports

+++ gc/src/main/java/com/bsgfb/cdp/gc/model/Node.java
@@ -4,12 +4,12 @@ import java.util.Arrays;
 import java.util.List;
 
 public class Node {
-    private long[] array;
+    private byte[] array;
     private List<Node> links;
 
     public Node(int byteSize) {
-        this.array = new long[byteSize];
-        Arrays.fill(this.array, Long.MAX_VALUE);
+        this.array = new byte[byteSize];
+        Arrays.fill(this.array, Byte.MAX_VALUE);
     }
 
     public List<Node> getLinks() {
